doc: add comments to opts
