+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 .DS_Store
 dist
+strapi-community-astro-loader-*.tgz
