+++ src/index.ts
@@ -1,1 +1,2 @@
 export { strapiLoader } from "./astro-loader.js";
+export { StrapiLoaderOptions } from "./options";
